% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_combn_fu.R
\name{auc_combn_fu}
\alias{auc_combn_fu}
\title{auc_combn_fu Function}
\usage{
auc_combn_fu(DATA, DV, IV, COMBN_VAL, ALGORITHM = "+")
}
\arguments{
\item{DATA:}{put data frame}

\item{DV:}{put Dependent Variable like this way: c("\emph{DV}")}

\item{IV:}{put Independent Variable like this way: c("\emph{IV1}", "\emph{IV2}", ...)}

\item{COMBN_VAL:}{consider how many combinations you want: 1, 2, 3, ...}

\item{ALGORITHM:}{you can choose "+" for multiplicity or "*" for interactions}
}
\description{
This function fu auc from combinations
}
\examples{
auc_combn_fu(mtcars,c("vs"),c("mpg","cyl","disp","hp","drat"),5)
auc_combn_fu(mtcars,c("vs"),c("mpg","cyl","disp","drat"),4)
auc_combn_fu(mtcars,c("vs"),c("mpg","cyl","disp","hp","drat","wt","qsec","am","gear","carb"),3)
}
\keyword{AUC}
\keyword{Combination}
\keyword{Follow-up}
