% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_combn_listup.R
\name{auc_combn_listup}
\alias{auc_combn_listup}
\title{auc_combn_listup Function}
\usage{
auc_combn_listup(
  DATA,
  DV,
  IV,
  IV.REF,
  COMBN_VAL = 1,
  ALGORITHM = "+",
  DISPLAY.NEVENT = FALSE,
  DISPLAY.EQUATION = FALSE,
  SEED = 123
)
}
\arguments{
\item{DATA:}{put data frame}

\item{DV:}{put Dependent Variable like this way: c("\emph{DV}")}

\item{IV:}{put Independent Variable like this way: c("\emph{IV1}", "\emph{IV2}", ...)}

\item{IV.REF:}{choose one of Reference Independent Varaible from IV}

\item{COMBN_VAL:}{consider how many combinations you want: 1, 2, 3, ...}

\item{ALGORITHM:}{you can choose "+" for multiplicity or "*" for interactions}

\item{DISPLAY.NEVENT:}{If TRUE then print EVENT/N, default is FALSE.}

\item{DISPLAY.EQUATION:}{If TRUE then print equation, default is FALSE.}

\item{SEED:}{set.seed, default is 123. for calculating bootstrap CI.}
}
\description{
This function can list-up AUC from multiple combinations.
}
\examples{
auc_combn_listup(mtcars,"vs",c("mpg","cyl","disp","hp","drat"),"mpg",1)
auc_combn_listup(mtcars,"vs",c("mpg","cyl","disp","hp","drat"),"mpg",2)
auc_combn_listup(mtcars,"vs",c("mpg","cyl","disp","hp","drat"),"mpg",3)
}
\keyword{AUC}
\keyword{Combination}
\keyword{Follow-up}
