% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delong_test.R
\name{delong_test}
\alias{delong_test}
\title{delong_test() Function}
\usage{
delong_test(x, y)
}
\arguments{
\item{x, }{y : two auc from pROC objects}
}
\description{
This function multiplies a and b
}
\examples{
delong_test()

}
\keyword{delong_test}
