% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_if_else.R
\name{ie}
\alias{ie}
\title{nested_if_else() Function}
\usage{
ie(...)
}
\arguments{
\item{if_stat, }{then}
}
\description{
source: https://www.r-bloggers.com/a-wrapper-around-nested-ifelse/
This function find formula from string
}
\examples{
ie(
  i(DATA$EDUC \%in\% c(1,2), 1),
  i(DATA$EDUC==3, 2),
  i(DATA$EDUC==4, 3),
  e(888)
)
}
\keyword{nested_if_else}
