% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit_f.R
\name{strsplit_f}
\alias{strsplit_f}
\title{strsplit_f() Function}
\usage{
strsplit_f(string, find)
}
\arguments{
\item{string, }{find: find formula from string}
}
\description{
This function find formula from string
}
\examples{
strsplit_f()

}
\keyword{strsplit_f}
