% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab1_r3.R
\name{tab1_r3}
\alias{tab1_r3}
\title{tab1_r3 function}
\usage{
tab1_r3(
  DATA,
  ID,
  TIME,
  GROUP,
  FUNC = "mean",
  CONDIGITS = 1,
  CATDIGITS = 1,
  TYPE = 1
)
}
\arguments{
\item{DATA:}{put data frame}

\item{ID:}{put id varaible}

\item{TIME:}{put repeated measures varaible (for stratified)}

\item{FUNC:}{"mean" or "median"}

\item{CONDIGITS:}{digits of continuous variable}

\item{CATDIGITS:}{digits of categorical variable}

\item{TYPE:}{type of print FUNC}
}
\description{
This function create table1 in biostatistics for repeated measures, v3: by time: count n(id)
}
\examples{
ddl <- pbcseq_modified
# baseline: count n(id)
tab1_r1(ddl, "id", "day")
# by overall: count n(case)
tab1_r2(ddl, "id", "day")
# by time: count n(id)
tab1_r3(ddl, "id", "day", "day.g")
}
\keyword{biostatistics}
\keyword{create}
\keyword{data}
\keyword{for}
\keyword{in}
\keyword{measures}
\keyword{repeated}
\keyword{table1}
