% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_combn_fu.R
\name{auc_combn_fu}
\alias{auc_combn_fu}
\title{auc_combn_fu Function}
\usage{
auc_combn_fu(DATA, DV, IV, COMBN_VAL, ALGORITHM = "+")
}
\arguments{
\item{DATA, }{DV, IV, COMBN_VAL, ALGORITHM}
}
\description{
This function fu auc from combinations
}
\examples{
auc_combn_fu(mtcars,c("vs"),c("mpg","cyl","disp","hp","drat"),5) \%>\% View()
auc_combn_fu(mtcars,c("vs"),c("mpg","cyl","disp","drat"),4) \%>\% View()
auc_combn_fu(mtcars,c("vs"),c("mpg","cyl","disp","hp","drat","wt","qsec","am","gear","carb"),3) \%>\% View()
}
\keyword{AUC}
\keyword{Combination}
\keyword{Follow-up}
