% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc_combn_listup.R
\name{auc_combn_listup}
\alias{auc_combn_listup}
\title{auc_combn_listup Function}
\usage{
auc_combn_listup(
  DATA,
  DV,
  IV,
  VAR0,
  COMBN_VAL = 1,
  ALGORITHM = "+",
  EQUATION = FALSE,
  SEED = 123
)
}
\arguments{
\item{DATA:}{put data frame}

\item{DV:}{put Dependent Variable like this way: c("\emph{DV}")}

\item{IV:}{put Independent Variable like this way: c("\emph{IV1}", "\emph{IV2}", ...)}

\item{COMBN_VAL:}{consider how many combinations you want: 1, 2, 3, ...}

\item{ALGORITHM:}{you can choose "+" for multiplicity or "*" for interactions}
}
\description{
This function can list-up AUC from multiple combinations.
}
\examples{
auc_combn_listup(mtcars,c("vs"),c("mpg","cyl","disp","hp","drat"),5)
auc_combn_listup(mtcars,c("vs"),c("mpg","cyl","disp","drat"),4)
auc_combn_listup(mtcars,c("vs"),c("mpg","cyl","disp","hp","drat","wt","qsec","am","gear","carb"),3)
}
\keyword{AUC}
\keyword{Combination}
\keyword{Follow-up}
