% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab1.R
\name{TAB1}
\alias{TAB1}
\title{table1 function}
\usage{
TAB1(
  DATA,
  STRATA = NULL,
  OVERALL = FALSE,
  CONDIGITS = 1,
  CATDIGITS = 1,
  FUNC = "mean",
  TYPE = 1,
  MARGIN = 2
)
}
\arguments{
\item{DATA:}{put data frame}

\item{STRATA:}{put stratification varaible}

\item{OVERALL:}{if you want overall statistics, set TRUE}

\item{CONDIGITS:}{digits of continuous variable}

\item{CATDIGITS:}{digits of categorical variable}

\item{FUNC:}{"mean" or "median"}

\item{TYPE:}{type of print FUNC}

\item{MARGIN:}{margin of percentage. 1 as row percent, and 2 as col percent}
}
\description{
This function create table1 in biostatistics
}
\examples{
ddd <- mtcars
ddd <- preprocess(ddd)
TAB1(ddd)
TAB1(ddd, STRATA = "cyl")
TAB1(ddd, STRATA = "cyl", OVERALL = TRUE)
TAB1(ddd, STRATA = "cyl", OVERALL = TRUE, CATDIGITS = 0)
TAB1(ddd, STRATA = "cyl", OVERALL = TRUE, CONDIGITS = 3, CATDIGITS = 0, FUNC = "median", TYPE = 1, MARGIN = 1)
}
\keyword{biostatistics}
\keyword{create}
\keyword{in}
\keyword{table1}
